fix: split health tests by layer — move startHealthServer to integration
health.unit.test.ts now contains only createHealthApp() tests (no I/O).
startHealthServer() tests that open real Bun.serve() sockets are moved
to health.integration.test.ts, per docs/test-readiness/test-system.md
unit-test layer contract.
